feat: renaming, some doc comments
