+++ lib/src/components/data-table/DataTableTable.tsx
@@ -3,10 +3,19 @@ import * as React from 'react'
 import { Table } from '../table'
 import { DataTable } from './DataTable'
 
-type DataTableTableProps = Omit<React.ComponentProps<typeof Table>, 'children'>
-export const DataTableTable: React.FC<DataTableTableProps> = (props) => (
+type DataTableTableProps = Omit<
+  React.ComponentProps<typeof Table>,
+  'children'
+> & {
+  sortable?: boolean
+}
+
+export const DataTableTable: React.FC<DataTableTableProps> = ({
+  sortable = true,
+  ...props
+}) => (
   <Table {...props}>
-    <DataTable.Header sortable />
+    <DataTable.Head sortable={sortable} />
     <DataTable.Body />
   </Table>
 )
